requirements.txt
# Minimal dependencies
openai
numpy
pytest
anthropic
python-dotenv